+++ src/backend/joanie/admin_urls.py
@@ -51,11 +51,16 @@ admin_router.register(
 admin_router.register("users", api_admin.UserViewSet, basename="admin_user")
 admin_router.register("orders", api_admin.OrderViewSet, basename="admin_orders")
 admin_router.register(
+    "batch-orders", api_admin.BatchOrderViewSet, basename="admin_batch_orders"
+)
+admin_router.register(
     "enrollments", api_admin.EnrollmentViewSet, basename="admin_enrollments"
 )
 admin_router.register("teachers", api_admin.TeacherViewSet, basename="admin_teachers")
 admin_router.register("skills", api_admin.SkillViewSet, basename="admin_skills")
-
+admin_router.register(
+    "discounts", api_admin.DiscountViewSet, basename="admin_discounts"
+)
 # Admin API routes nested under a course
 admin_course_related_router = DefaultRouter()
 admin_course_related_router.register(