+++ src/backend/joanie/core/models/contracts.py
@@ -93,8 +93,9 @@ class Contract(BaseModel):
         verbose_name=_("order"),
         on_delete=models.PROTECT,
         editable=False,
+        null=True,
+        blank=True,
     )
-
     # Set on contract generation
     definition_checksum = models.CharField(
         max_length=255, editable=False, blank=True, null=True
@@ -239,7 +240,8 @@ class Contract(BaseModel):
         self.definition_checksum = checksum
         self.signature_backend_reference = reference
         self.save()
-        self.order.flow.update()
+        if self.order:
+            self.order.flow.update()
 
     def reset_submission_for_signature(self):
         """
@@ -251,7 +253,8 @@ class Contract(BaseModel):
         self.definition_checksum = None
         self.signature_backend_reference = None
         self.save()
-        self.order.flow.update()
+        if self.order:
+            self.order.flow.update()
 
     def is_eligible_for_signing(self):
         """