+++ src/backend/joanie/tests/core/api/admin/batch_order/test_update.py
@@ -1,0 +1,50 @@
+"""Test suite for the admin batch orders API update endpoint."""
+
+from http import HTTPStatus
+
+from django.test import TestCase
+
+from joanie.core import factories
+
+
+class BatchOrderAdminApiUpdateTestCase(TestCase):
+    """Test suite for the admin batch orders API update endpoint."""
+
+    def test_api_admin_batch_orders_update_anonymous(self):
+        """Anonymous user should not be able to update a batch order."""
+
+        batch_order = factories.BatchOrderFactory()
+
+        response = self.client.put(
+            f"/api/v1.0/admin/batch-orders/{batch_order.id}/",
+        )
+
+        self.assertEqual(response.status_code, HTTPStatus.UNAUTHORIZED, response.json())
+
+    def test_api_admin_batch_orders_update_authenticated(self):
+        """Authenticated user should not be able to update a batch order."""
+        user = factories.UserFactory(is_staff=False, is_superuser=False)
+        self.client.login(username=user.username, password="password")
+
+        batch_order = factories.BatchOrderFactory()
+
+        response = self.client.put(
+            f"/api/v1.0/admin/batch-orders/{batch_order.id}/",
+        )
+
+        self.assertEqual(response.status_code, HTTPStatus.FORBIDDEN, response.json())
+
+    def test_api_admin_batch_orders_update_authenticated_admin(self):
+        """Authenticated admin user should not be able to update a batch order."""
+        user = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=user.username, password="password")
+
+        batch_order = factories.BatchOrderFactory()
+
+        response = self.client.put(
+            f"/api/v1.0/admin/batch-orders/{batch_order.id}/",
+        )
+
+        self.assertEqual(
+            response.status_code, HTTPStatus.METHOD_NOT_ALLOWED, response.json()
+        )